Add files via upload
Data Description:
A sample survey data (Sample Survey.xlsx) is provided for analysis. Samples were collected
over a period of nine days where the respondents were asked questions related to elections.
Survey weights are in the “weight” column. The description for each column is provided in the
excel sheet under tab named “Description”. Applicants are requested to use python-pandas for
all data manipulations and matplotlib for any plotting related questions. Applicants are expected
to not use for-loops or any kind of iterations for solving the questions. Use in built pandas
functions only.